+++ test/unittests/engine/midi_dispatcher_test.cpp
@@ -16,6 +16,8 @@ using namespace sushi::internal;
 using namespace sushi::internal::engine;
 using namespace sushi::internal::midi_dispatcher;
 
+constexpr float TEST_SAMPLE_RATE = 48000.0;
+
 const MidiDataByte TEST_NOTE_ON_CH2   = {0x91, 62, 55, 0}; /* Channel 2 */
 const MidiDataByte TEST_NOTE_OFF_CH3  = {0x82, 60, 45, 0}; /* Channel 3 */
 const MidiDataByte TEST_CTRL_CH_CH4_67 = {0xB3, 67, 75, 0}; /* Channel 4, cc 67 */
@@ -119,8 +121,8 @@ protected:
         _module_under_test.set_frontend(&_mock_frontend);
     }
 
-    EngineMockup _test_engine{48000};
     EventDispatcherMockup _test_dispatcher;
+    EngineMockup          _test_engine {TEST_SAMPLE_RATE};
     ::testing::NiceMock<MockMidiFrontend> _mock_frontend{nullptr};
     MidiDispatcher _module_under_test{&_test_dispatcher};
 };