+++ HISTORY.md
@@ -1,3 +1,14 @@
+## Next
+
+New Features:
+  * C++20 support
+  * Updated Elklog
+
+Fixes:
+  * Improved handling of audio pause / resume
+  * Xrun detection in audio frontends
+
+
 ## 1.1.0
 
 New features: